Update bcast test roots and fix psync sizes
The broadcast tests have an item that expect well-defined results when
the root PE is different across the active set, which is not permitted
by the OpenSHMEM specifcation.  Also, broadcast tests should use
SHMEM_BCAST_SYNC_SIZE instead of SHMEM_COLLECT_SYNC_SIZE.  Finally, the
wait for 5 sleep cycles after each test item is unnecessary, because the
buffer should be ready for re-use after the call to broadcast returns.

Signed-off-by: David Ozog <david.m.ozog@intel.com>